+++ esp-hal/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - A new default feature `exception-handler` was added (#3887)
 - `AesBackend, AesContext`: Work-queue based AES driver (#3880)
 - `aes::cipher_modes`, `aes::CipherModeState` for constructing `AesContext`s (#3895)
+- `RsaBackend, RsaContext`: Work-queue based RSA driver (#3910)
 
 ### Changed